requirements-dev.txt
asyncio==3.4.3
black==22.10.0
celery==5.2.7
circuitbreaker==1.4.0
colorclass==2.2.2
django-enumfield==3.0
django-filter==22.1
django-health-check==3.17.0
django-prometheus==2.2.0
drf-spectacular==0.24.2
faust-streaming==0.10.0
flake8==5.0.4
flake8-import-order==0.18.1
install==1.3.5
ipython==8.6.0
isort==5.10.1
mode==4.3.2
mysqlclient==2.1.1
pip-chill==1.0.1
pre-commit==2.20.0
pymemcache==4.0.0
pytest==7.2.0
python-consul2==0.1.5
python-dotenv==0.21.0
redis==4.4.0
robinhood-aiokafka==1.1.6
stripe==4.2.0
websockets==10.4
